Twelveth problem
Array that contains integer values, some of these values are repeated with an even number of repetitions, and only one value is repeated with an odd number of repetitions. Write a C function that’s take the array as input and the array size and return the number which has odd numbers of repetitions.